Add tasks for release monitoring and getting built images from Konflux (#116)
Add two scripts:
- Getting Konflux operator catalog images
- Monitoring the release. 
I think this might simplify interacting with Konflux UI and could
prevent releaser engineers opening many tabs and monitor each of them
individually.

Also tune README a bit

*NOTE:* `pipelinerun` CR TTL is about 5 days. So getting build older
than 5 days is an issue with this command and UI has to be used instead.
I added a message which sugget to use UI if images are not found.

---------

Co-authored-by: Tom Martensen <tmartens@redhat.com>
Co-authored-by: Misha Sugakov <537715+msugakov@users.noreply.github.com>